requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
moviepy==1.0.3
openai-whisper==20231117
ffmpeg-python==0.2.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
python-multipart==0.0.9
